Revert "fixing reflection constructor to use int instead of long for ksessionId"
This reverts commit 2e5e7bcd356b42886fb09c0586f277d56464b95e.

+++ drools-persistence-jpa/src/main/java/org/drools/persistence/jpa/KnowledgeStoreServiceImpl.java
@@ -81,7 +81,7 @@ public class KnowledgeStoreServiceImpl
 
         try {
             Class< ? extends CommandExecutor> serviceClass = getCommandServiceClass();
-            Constructor< ? extends CommandExecutor> constructor = serviceClass.getConstructor( long.class,
+            Constructor< ? extends CommandExecutor> constructor = serviceClass.getConstructor( int.class,
                                                                                               KnowledgeBase.class,
                                                                                               KnowledgeSessionConfiguration.class,
                                                                                               Environment.class );
